+++ lib/features/wallet/wallet_model.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:vase/constants.dart';
+import 'package:vase/config/constants.dart';
 import 'package:vase/lotus/lotus.dart';
 import 'package:vase/wallet/keys.dart';
 import 'package:vase/wallet/wallet.dart';
